v0.0.7 support control transfer
1. Support save settings
2. Protect mouse and protect keyboard by default
3. Support control transfer
4. Fix thread safety issue of logger

+++ src/lib/usb/usbconfigurationdescriptor.h
@@ -127,7 +127,7 @@ namespace usb {
          * the index of interface, must less than bNumInterfaces.
          * @return the specified interface
          */
-        UsbInterface *interface(int index) const;
+        UsbInterface *interface(uint8_t index) const;
 
         /**
          * @brief device
